fix: /plot command not work quite right.

+++ src/main/java/com/ixbob/plotplugin/command/CommandPlot.java
@@ -71,8 +71,8 @@ public class CommandPlot implements CommandExecutor {
             }
 
             DBObject playerData = Utils.getPlayerData(uuid);
-            double x = (double) playerData.get("X_from") - 50;
-            double z = (double) playerData.get("Z_from") - 50;
+            double x = ((Integer) playerData.get("X_from")).doubleValue() - 50;
+            double z = ((Integer) playerData.get("Z_from")).doubleValue() - 50;
             player.teleport(new Location(player.getWorld(), x, 52, z));
             player.sendTitle(LangLoader.get("plot_welcome_back_title"),
                     String.format(String.format(LangLoader.get("plot_owner_subtitle"), player.getName())), 10, 60, 10);
